[ARRISAPOL-2261] Report buffering level based on brcmvidfilter reporting
MediaPlayerPrivateGStreamer starts playback when buffering level reaches
100% the first time. Buffering level is computed only based on GstQueue2
reporting. This is wrong, because there are more queues further in the
pipeline, which drain GstQueue2 out of data as soon as they can. This
causes slow start of progresive videos, especially on slow connections.

This patch modifies this algorithm and takes into account how much data
was buffered in brcmvidfilter.

Based on this upstream patch:
https://github.com/WebPlatformForEmbedded/WPEWebKit/commit/14f712084b349ab70e87352d73dc02ee42c1a99e